+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "test-app",
-  "version": "0.1.0",
+  "name": "CustomCodeEditor",
+  "version": "stable",
   "private": true,
   "dependencies": {
     "@fortawesome/fontawesome-free": "^6.5.2",
